[project @ 2002-04-15 10:23:53 by as49]
Realized the stupid idea to take away the first argument of signal handlers.
In C it is only possible to omit the last arguments which we already made use
of by ignoring the user pointer. The sadly means that for Solaris we are
restricted to 3 arguments in the callback (the object pointer + 3 word args).

Fixed a dependency bug with another bad solution.

darcs-hash:20020415102355-d90cf-b9aef03b7f44acf1496f6acf75fc569e9069c394.gz