[project @ 2002-04-16 13:19:51 by as49]
Add a new goal to make a source distribution. make dist now uses the verion
in VERSION to build a tarball consisting of the CVS sources plus the
configure shell script. The configure skript does not always seem to work
when using it on other platforms. Changed the installations directions and
the TODO list to reflect the Solaris fix, the built-in c2hs and this add-on.

darcs-hash:20020416131954-d90cf-21c75f1f5723ae42df5880a36df959c1ef5cb698.gz