[project @ 2003-07-09 21:12:27 by as49]
Made c2hs generate code that is compatible with GHC 6.00 (or the "new"
FFI). GHC does not accept ForeignPtr as arguments anymore. Each call
to a function which takes such an argument is now marshaled as follows:
Instead of inserting just the function name, a lambda expression is
emitted which strips off the newtype constructor if one was generated
by c2hs. The resulting ForeignPtr is then converted to a pointer via
withForeignPtr. For this to work, the constructor of the specific data
type has to be in scope as well as the withForeignPtr function.

darcs-hash:20030709211227-d90cf-d0763465acdc23de4fb4b207efd4b4b120988439.gz