[project @ 2004-12-18 20:46:29 by duncan_coutts]
make chsDepend understand hierarchical modules names and change the dependency
of .hs files to be on the .chi files rather than the .chs files. This is how it
is done for .o and .hs files, that is, the .o file depends on the .hi files.

darcs-hash:20041218204629-d6ff7-708bde69e84bf613522b8a1286b4626d66adb551.gz