[project @ 2005-03-25 19:14:35 by duncan_coutts]
Various changes to the code generator:

ModuleScan.hs: match up against existing .hs modules if there is no .chs or
.chs.pp files.

ApiGen.hs: minor knock-on change from above.

MarshalFixup.hs: more fixup info.

CodeGen.hs: pass whether the function is a constructor on to the marshaling
functions. This allows them to add castPtr for constructors. Output the
constructors in the same order as they occur in the original module (same as
we do for functions). Change KnownSymbols map to include the full chaing of
parents of each class rather than just a clasification of whether they inherit
from GObject or GtkObject. For leaf classes, generate the properties with the
specific type rather than it's corresponding typeclass.

Marshal.hs: The KnownSymbols map change described above. Add castPtr for
constructors and include the specific types that we are casting from and to,
just for safty in case the code generator gets it wrong and disagrees with
c2hs on types. Add marshaling for gunichar.

Makefile: add gobject-api.xml to the --includeapi list so we don't need a
special case for GObject in the code generator.

glib-sources.xml: make this into glib not gobject.

gobject-sources.xml: add gobject seperately.

darcs-hash:20050325191455-d6ff7-fbe28dabf971ff7640a2464fbeb24f321ac962d2.gz