Allow GC-managed allocation of cairo image surfaces
Make surfaces use a ForeignPtr and make a createImageSurface variant that
uses a finaliser on the ForeignPtr to allow the GC to manage the lifetime of
the surface. The surfaceFinish function can be used to manually release the
memory used by the surface. So guaranteed memory management for surfaces is
still possible.
Also export surfaceFlush and surfaceFinish.

darcs-hash:20060125154424-b4c10-fb0d51d104f6de5b7e693092f9494389533cdb1f.gz