Update cairo clock demo
It now makes use of the new createImageSurface function to reallocate the
cached image layers when the window resizes. This means it now resizes
properly and with no GC-induced leakage because we use surfaceFinish to
manually release the image data associated with the surface.
Also improve the low quality redraw mode to include the clock frame.
The only things missing now are fixing the aspect ratio and making it work
with a colormap that has an alpha channel.

darcs-hash:20060125162628-b4c10-0c16cec657495162f67c1619c2dcabc17c25bf27.gz