Change textBufferNew to take a TextTagTable, not just any subclass
in 0.9.10 it took a TextTagTable. This was changed to any instance of
TextTagTableClass. But actually this is not a correct generalisation because
we cannot construct a TextBuffer with a SourceTagTable, only a TextTagTable.
Also it causes a type class ambiguity when one passes Nothing. One would have
to give a type annotation like:
textbuffer <- textBufferNew (Nothing :: Maybe TextTagTable)
So I'm reverting this so it just takes a TextTagTable.

darcs-hash:20060129210816-b4c10-4faf9380b152e2a8f38611d1b182bf75ab82f0f8.gz