Use the flush command rather than iterating the main loop
This has the same effect of actualy sending the command to the X11 server to
close the window.
We should probably use flush in the implementation of mainGUI so that users
need not be concerned with it.

darcs-hash:20060201095605-b4c10-cc0e043f0364b2a03d73a3da32e3429cf9cd2137.gz