More fixes for the treelist widgets.
This patch incorporates some of the fixes necessary to make the new
custom store available to the widgets that use the treelist stores.
It is far from complete. For a typed set of widgets some of these changes
need to be changed (and partly reverted perhaps).

darcs-hash:20060319184139-91c87-370c4233a33a84e566adf175cb2d168e1c0a75e7.gz