Workaround for font enumeration bug in Pango.
This patch removes the cairoFontMapNew function and adds
cairoFontMapGetDefault. While the former function can be bound and
is itself correct, certain functions like pangoFontMapListFamilies
are not. The latter function creates FontFace structures (and others)
whose reference is added to the FontMap object. The FontFace
objects themselves contain a pointer to
the FontMap that they are embedded in. Alas, they do not add a reference
to the FontMap they have locally in their own structure.
Hence, when Haskell garbage collects the FontMap created by cairoFontMapNew
the FontMap that contains the FontFaces vanishes and the FontFace (which
itself is a GObject) stays alive and has thereby a dangling pointer to
the FontMap. This is certainly a bug in Pango but they might not want to
fix this since it involves breaking cycles in reference counting.

darcs-hash:20060502183537-91c87-a04d133953530448c93713c9ba3d155b564ea253.gz