Add support of new floating reference in GObject.
The floating reference needs to be ignored just as in GtkObject.
There are two issues with this patch:
a) it's untested, it might not even compile
b) it should use GTK_CHECK_VERSION when it should use GLIB_CHECK_VERSION
For b) we don't have the appropriate macros yet, and this change will
probably stay the only thing that needs this macro.

darcs-hash:20060502184512-91c87-1b8399a1aae349a36fc7aad6a3314de60784c256.gz