More extensive scanning of source modules
Instead of just looking for c2hs call hooks we now scan the source
.hs/.chs modules for function definitions, and save the complete code
and documentation for each function. From the function body we
extract any c2hs call hooks. So now gnerated functions are ordered by
the possition of the corresponding functions in the scanned module
rather than by the position of the corresponding C call.
The former is typically more accurate.
We want the complete docs and code for functions so that we can
implement the feature where if the hashes match that we emit the
original scanned code/docs rather than the generated stuff.

darcs-hash:20061114044629-b4c10-0ea1d060aea4014744ea680934763a9ecc2c7a39.gz