Preserve original code when the generated code matches the hash
For both the code and docs we now emit the original code and/or docs
for a function if the original module had a hash which matches that
of the generated code / docs. We also emit the hash if the code/docs
do not match the original.
So the point is that you can merge the hash for the docs and/or code
into the hand-written module and then next time the code gen is run,
it'll spot that you are happy with the differences between the
hand written and generated code and emit the orignal code rather than
the generated code. So it massively reduces the sizes of the diffs
between the orignal and generated modules. Though if the generated code
changes at all for any reason, then the code gen will still emit the
generated code/docs so that you can check again that you're happy with
the differences.
This might happen if the code gen is updated to emit different code, or
if the gtk docs get updated.

darcs-hash:20061115052030-b4c10-2d2240289750d6d1a792030b238157627ac6da25.gz