Change the new TreeStore api to be more consistent
To be more consistent with the api of the old tree store and also
more consistent with the new list store. The naming convention is:
Get/SetValue does just the label of the node not the subtrees too.
*Tree suffix does a whole subtree, eg: treeStoreGetTree
For completeness we should probably have treeStoreSetTree.
Also fix treeStoreInsertTree to notify the view of all subnodes that
got inserted. This is necessary or the view does not display them.

darcs-hash:20061224010457-b4c10-03fd74484ab580ea2f5d871044ed80ba57cda92a.gz