Fix memory management of TextIter type.
The TextIter type is allocated and owned by the caller.
For Haskell we can use mallocForeignPtrBytes which allocates the memory
on the Haskell heap and automatically manages the lifetime via the GC.
Thanks to Jean-Philippe Bernardy for the nice bug report and test case.

darcs-hash:20061228192048-b4c10-e0b8fc46353040191828810db82f3adacf58a177.gz