Add traffic simulation demo by Maurício Antunes
This is a facinating demo, it simulates cars driving round a circut
you can use the mouse to tempoarily block the cars causing a jam
and see if it remains a traffic jam or can recover. Wether or not it
recovers depends on the traffic density (which is adjustable).

darcs-hash:20070108134629-b4c10-0a292550f3337d01d95f61079b2078f446dc910f.gz