Allow use of the threaded RTS
initGUI still fails when used with the threaded RTS, though only when
used in a stand-alone program. When used in GHCi it doesn't complain
on the theory that people don't actually use threads much in GHCi,
where as they have to go out of their way to link a stand-alone
program with -threaded, so they must be intending to use OS threads.
The error message refers to a new function: unsafeInitGUIForThreadedRTS
as you might expect this doesn't fail when used with the threaded RTS.
Also added a couple functions: postGUISync and postGUIAsync to make
it easier to use GUI functions from other threads.
unsafeInitGUIForThreadedRTS now needs to initialise glib's thread
support which is a little tricky since it's not allowed to be
initialised more than once in a program run. This required some C code,
since the functionality is only provided as a macro.

darcs-hash:20070129002739-b4c10-3b08d5467bbbfbeeb64770828634c1ea9069c37d.gz