A frist go at a comprehensive tutorial.
This document consists of an introduction which I wrote yesterday evening and
a section on the new tree/list model which is quite old and has been lying
around on my laptop for quite a while. The code of the tree/list stuff
is not tested. Furthermore, the screenshot is missing.

darcs-hash:20070129093903-e1dd6-b08a3c9a45058b39056803298cb3bad57b749617.gz