Use a different trick to ensure we call g_thread_init at most once
The glib function "g_thread_init" aborts the program if you call it
more than once. The glib docs suggest a particular macro to test if
it's already been initialised. Obviously we can't use the macro in
Haskell so I added a bit of C glue code to do that bit. Sadly, the
way that macro works is to check a C global variable. In windows,
importing global variables across .dll boundaries is rather tricky.
The gnu linker gets around it with some cunning hackery, sadly GHCi
doesn't have the same cunning and so it can't link HSgtk.o if we have
a reference to this glib global var. So we need a different trick.
So now it works using a Haskell global var to track wether or not
we've initialised the glib threads module. Yes this uses the evil
top-level IORef trick. It's not nice but it works, even in Windows.

darcs-hash:20070130173701-b4c10-fff4a626559fd14994d7690f1fca85c2de62a516.gz