Use the right cairo backend headers in the package registration files
If we're using one of the backend headers then client code might need it too.
So we have to be accurate about which ones we include. We can't inlcude ones
that are not enabled either or we'll get cpp errors.

darcs-hash:20070523095906-adfee-34cf0866673a94f49c3fdf41856d22f7fe5a6ea9.gz