Make the check for the GThread system being initialised more robust
We previously used an evil top level IORef however in GHCi that does
not work because that state gets reverted when we :reload, and so we
end up calling g_thread_init again which aborts.
So now we have a little C wrapper around the g_thread_supported macro
and we use that to check if g_thread_init was already done.

darcs-hash:20070630000326-adfee-593d2a54df72f936f7a14f68234ad79619657045.gz