Reimplement the way we keep track of gthread initialisation state
Sadly we cannot just call g_thread_supported() because this macro expands
into a reference to a gloabl variable that lives in the gthread dynamic lib.
On windows, GHCi's dynamic linker cannot cope with references to global vars
from C dlls. So instead we keep a global var in the wrapper, so it's in the
same package, rather than crossing a dll boundary. This allows us to track
the state, and because the state is allocated in C land, it survives :reload
in GHCi, which was the problem with the original scheme. Ugg.

darcs-hash:20070704012435-adfee-eab7ed081df1eca547841583b24e2f1fd04e6502.gz