Add columns to our Haskell tree model.
This patch adds the ability to lookup values from TreeStore and ListStore in terms of columns. Some information in e.g. ComboBox have properties that are not shown by CellRenderers directly. These properties can therefore not be set by using function in CellLayout and the only way to connect them to the model is by pretending they access a certain column in the model. Hence this patch adds the ability to access a Haskell model using a column number. The idea is that these column numbers are opaque to the user of Gtk2Hs. Functions that use columns are called widgetSetBlahSource where wiget is the widget and Blah is the property.

darcs-hash:20070715162232-e1dd6-9b86653b41fa4aa2ab3c4fc4b9319e6231775a87.gz