glib: add toFlags': like toFlags but works doesn't fail when we haven't defined toEnum for a bit
This is needed for a few reasons, the most important of which is for forward compatibility.
If a newer version of the upstream package defines a new flag for a flag type, we shouldn't
fail just because we don't recognize it. This version allows overlapping flags, for example, given:

data OpenFlags

darcs-hash:20071105150631-21862-4f8d7236a8ce2716ba65988c1c101cb216215b5a.gz