Use precise package versions when building and in package registration
Previously we were being sloppy and allowing any version in the registration
so there was no guarantee that the version we built against was the same as
the version that would get used when the package was installed. Similarly
we could build against packages that were registered per-user which would
then fail if we did a global install. Now it only looks for global packages
unless you ./configure --with-user-pkgconf
This version of the patch for gtk2hs HEAD also covers gstreamer and gnomevfs

darcs-hash:20071107133455-adfee-3f73f23fde9e4bfe7edd6ad297e759f31bb0af48.gz