Complete the default drag and drop interface for ListStore.
This patch completes the former patch which adds the C functions to the C custom model. With this patch it is possible to reorder the rows of a list store using drag and drop. The drag and drop for tree stores doesn't quite work yet since I'm not sure what the exact semantics are (do we move a row at a time, or a full tree? is the last index of the drop path the position where we insert the new row?). This patch also enables the reordering in the two listtest and treetest demo programs.

darcs-hash:20071202211958-e1dd6-582ecc4cdb001faebcd955caba84b17e6db36d07.gz