Fix stamp problems in tree implementation.
The story of TreeIter stamps is as follows: ListStore doesn't use them since
the iterators remain valid throughout. TreeStore needs to invalidate the
iterators each time a node is inserted since a new node might mean that it
can't be indexed anymore with the available number of bits at that level.
The function that inserts nodes correctly invalidated the tree iterators.
However, this function also calls methods in TreeModel, passing in order
to notify itself and the view. The passed in TreeIters have no stamp set,
thus a warning is raised as soon as these tree iters get back to the model,
usually in a form of ref'ing the node. This was slightly confusing as
setting the stamp was so far automatically done in the C interface
Gtk2HsStore.c. The stamp is now set explicity in the TreeIters before they
are passed to the functions in the model.

darcs-hash:20071204212616-e1dd6-70f4d8d2e7caaa3296cb1568be53c0831c38dba7.gz