Add the TreeSortable interface.
This interface allows a view to sort the underlying model. Since our
CustomStore does not implement this interface, the only object that
actually does implement it is the TreeModelSort proxy which takes a child
model and sorts its rows on-the-fly.

darcs-hash:20080103190026-e1dd6-1ecca3f64331cd61bd0c351531f07f2f4c377107.gz