Review the support for creating columns.
Column numbers must now be declared as constants. The extraction function
has then to be set by 'treeModelSetColumn' which replaces the function
to update a column. This makes everything more low-level and can lead
to type errors, however, the result is that Gtk will create warnings
if the requested value is of the wrong type when reading from the store.

darcs-hash:20080210191616-e1dd6-149bd0995f36faa8d8ed354420d82104f0606d7e.gz