Add a treeModelValueGet function.
This patch makes it possible to retrieve the content of a store using the
TreeModel interface, that is, by using predefined columns. The ColumnId
type had to be amended for that. I used the opportunity to move its
definition to the TreeModel module, where it belongs. As a consequence
the two functions in CustomStore whose name started with treeModel
have been renamed to customStoreGetRow and customStoreSetColumn since they
really only make sense on implementations of CustomStores. The original
names are no longer documented but still exported in order not to break
exisiting programs.

darcs-hash:20080915211739-e1dd6-4399617e60c47bb9a4a02a6c22e77d52d001d1a2.gz