Add new way to access information in events.
This patch makes it possible to pass around the pointer to an event structure
without marshalling all its content. The new modules is called EventM and
will replace the Events module. For now, I have stopped exporting Events
from Gtk by default. It might be best to force people to explicitly import
the events module from now on which might be ok, since it is rather low
level stuff that is only needed when a special widget is to be implemented.

The advantage of the monad encapsulation is that it is now possible to bind
functions that take a pointer to an event as parameter. There are only a few
of these, but it is very hard to create a pointer to an event from scratch.

darcs-hash:20081012222705-e1dd6-f53b205707b80a21927d327fba4fee6db0f13f5a.gz