Export on and after by default.
Since this release will break programs, I think it is the rigth time to
always export these two functions. There is a clash with a function in
Data.Function. I think this is not so serious since Data.Function does
not contain a lot of terribly useful functions and is thus not likely to
be imported by default.

darcs-hash:20081026163803-5ff50-0852e99728bad2f4a150f8d884dc7cbdfb83d74f.gz